refactor(timer): split timer_page into per-widget files
timer/widgets/:
- countdown_dial.dart           CountdownDial (the circular progress + MM:SS)
- preset_chips.dart             PresetChips (composer for the two rows + Custom)
- preset_chip_row.dart          PresetChipRow (a single Wrap row of ChoiceChips)
- timer_controls.dart           TimerControls (Start/Pause + Reset)
- alarm_done_dialog.dart        AlarmDoneDialog ("Rest over / Time to lift")
- custom_duration_dialog.dart   CustomDurationDialog (minutes + seconds input)

Page file shrinks from ~405 lines to ~155 lines.
